telegram_metatrader_app/requirements.txt
python-telegram-bot
MetaTrader5
PyQt5
SQLAlchemy
